+++ src/layers/domain/core/device/v1.py
@@ -175,6 +175,8 @@ class QuestionnaireResponseUpdatedEvent(Event):
 @dataclass(kw_only=True, slots=True)
 class DeviceReferenceDataIdAddedEvent(Event):
     id: str
+    keys: list[DeviceKey]
+    tags: list[str]
     device_reference_data: dict[str, list[str]]
     updated_on: str = None
 
@@ -201,15 +203,16 @@ class DeviceTag(BaseModel):
         if initialised_with_root and isinstance(item_to_process, str):
             _components = ((k, v) for k, (v,) in parse_qs(item_to_process).items())
         # Case 2: query components are provided (__root__=("foo", "bar"))
-        elif initialised_with_root:
+        elif initialised_with_root and not isinstance(item_to_process, dict):
             _components = ((k, v) for k, v in item_to_process)
-        # Case 3: query components are directly provided (("foo", "bar"))
+        # Case 3: query components are directly provided (("foo", "bar")) or (__root__=("foo", "bar"))
         else:
             _components = ((k, str(v)) for k, v in item_to_process.items())
 
         case_insensitive_components = tuple(
             sorted((k, v.lower()) for k, v in _components)
         )
+
         return {"__root__": case_insensitive_components}
 
     def dict(self, *args, **kwargs):
@@ -319,7 +322,7 @@ class Device(AggregateRoot):
         self.keys.remove(device_key)
         return DeviceKeyDeletedEvent(
             deleted_key=device_key.dict(),
-            id=self.id,
+            id=str(self.id),
             keys=[k.dict() for k in self.keys],
             tags=[t.value for t in self.tags],
         )
@@ -400,6 +403,8 @@ class Device(AggregateRoot):
         device_data = self.state()
         return DeviceReferenceDataIdAddedEvent(
             id=device_data["id"],
+            keys=device_data["keys"],
+            tags=[t.value for t in self.tags],
             device_reference_data=device_data["device_reference_data"],
         )