+++ src/layers/domain/repository/device_reference_data_repository/tests/test_device_reference_data_repository_v1.py
@@ -1,5 +1,3 @@
-import time
-
 import pytest
 from domain.core.device_reference_data import DeviceReferenceData
 from domain.core.enum import Environment
@@ -17,7 +15,6 @@ def test__cpm_device_reference_data_repository(
     repository: DeviceReferenceDataRepository,
 ):
     repository.write(device_reference_data)
-    time.sleep(1)
     result = repository.read(
         product_team_id=device_reference_data.product_team_id,
         product_id=device_reference_data.product_id,
@@ -33,7 +30,6 @@ def test__cpm_device_reference_data_repository_already_exists(
     repository: DeviceReferenceDataRepository,
 ):
     repository.write(device_reference_data)
-    time.sleep(1)
     with pytest.raises(AlreadyExistsError):
         repository.write(device_reference_data)
 
@@ -59,7 +55,6 @@ def test__cpm_product_repository_local(
     repository: DeviceReferenceDataRepository,
 ):
     repository.write(device_reference_data)
-    time.sleep(1)
     result = repository.read(
         product_team_id=device_reference_data.product_team_id,
         product_id=device_reference_data.product_id,
@@ -109,7 +104,6 @@ def test__cpm_device_reference_data_repository__search_not_empty(
     repository: DeviceReferenceDataRepository,
 ):
     repository.write(device_reference_data)
-    time.sleep(1)
     results = repository.search(
         product_team_id=device_reference_data.product_team_id,
         product_id=device_reference_data.product_id,