+++ src/layers/sds/epr/tests/test_getters.py
@@ -68,7 +68,7 @@ def test_get_mhs_device_data(mhs_1: NhsMhs):
         "Date DNS Approved": "yesterday",
         "Date Requested": "a week ago",
         "MHS FQDN": "my-fqdn",
-        "MHS Party key": "AAA-123456",
+        "MHS Party Key": "AAA-123456",
         "Managing Organization": "AAA",
         "Product Name": "My EPR Product",
         "Requestor URP": "requester-123",
@@ -99,9 +99,12 @@ def test_get_as_data(accredited_system_1: NhsAccreditedSystem):
         "Date Approved": "today",
         "Date Requested": "a week ago",
         "ODS Code": "AAA",
-        "Party Key": "AAA-123456",
+        "MHS Party Key": "AAA-123456",
         "Product Key": "key-123",
         "Requestor URP": "requester-123",
+        "Product Name": None,
+        "Product Version": None,
+        "Temp UID": None,
     }
     assert as_data.created_on.date() == now().date()